app/requirements.txt
uvicorn
fastapi
python-keycloak
